pyproject.toml
[project]
name = "ortler"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    "openreview-py",
    "argcomplete",
    "numpy",
    "openai",
    "pdf2image",
    "reportlab",
    "python-dotenv",
    "termcolor",
]

[project.scripts]
ortler = "ortler.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
